+++ src/main/java/model/livings/Giraffe.java
@@ -3,7 +3,7 @@ package model.livings;
 public class Giraffe extends Herbivore{
 
 
-    public Giraffe(int x, int y, double health, double hunger, int speed) {
+    public Giraffe(int x, int y, int health, int hunger, int speed) {
         super(x, y, health, hunger, speed);
     }
