+++ src/main/java/model/livings/Lion.java
@@ -3,7 +3,7 @@ package model.livings;
 public class Lion extends Carnivore {
 
 
-    public Lion(int x, int y, double health, double hunger, int speed, double attackValue) {
+    public Lion(int x, int y, int health, int hunger, int speed, int attackValue) {
         super(x, y, health, hunger, speed, attackValue);
     }
