+++ src/main/java/model/livings/Snake.java
@@ -2,7 +2,7 @@ package model.livings;
 
 public class Snake extends  Carnivore{
 
-    public Snake(int x, int y, double health, double hunger, int speed, double attackValue) {
+    public Snake(int x, int y, int health, int hunger, int speed, int attackValue) {
         super(x, y, health, hunger, speed, attackValue);
     }
