+++ src/main/java/model/livings/Zebra.java
@@ -2,7 +2,7 @@ package model.livings;
 
 public class Zebra extends Herbivore{
 
-    public Zebra(int x, int y, double health, double hunger, int speed) {
+    public Zebra(int x, int y, int health, int hunger, int speed) {
         super(x, y, health, hunger, speed);
     }
